tests(reports) extract mock_reports_server function to helpers
